requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.4
coverage==6.0.2
pytz==2021.1
python-dotenv==0.19.1
sqlparse==0.4.1
psycopg2-binary
celery==5.0.5
redis==3.5.3
requests
